Rewrite the command line flags as commands.
Previously these were written as flags. This leads to the slightly weird behavior of rendering the update graph at the same time as the render graph. While this *can* be useful, it's a little confusing and hard to extend. For example, if we want to extend to using per-command flags (e.g., filter by system set), it's not clear this flag doesn't work during `--dump-render`.