Rename dump-update-schedule to dump-schedule and allow dumping any schedule.
This just looks for the schedule with the same name and dumps it instead of just the Update schedule.

Sadly this does leave us to some ambiguity. We differentiate schedules by their `Debug` implementation, which is not guaranteed to be unique. However, it's very unlikely that users are going to have two of the same schedule name in the same Bevy app, so we can just ignore this.